requirements.txt
absl-py==1.0.0
aniso8601==9.0.1
appdirs==1.4.4
asttokens==2.0.5
attrs==21.4.0
audioread==2.1.9
backcall==0.2.0
brotlipy==0.7.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.2
coloredlogs==14.0
cycler==0.11.0
Cython==0.29.28
decorator==5.1.1
Distance==0.1.3
dnspython==2.2.1
dpcpp-cpp-rt==2022.1.0
editdistance==0.6.0
et-xmlfile==1.1.0
eventlet==0.33.0
executing==0.8.3
ffmpeg==1.4
filelock==3.6.0
Flask==2.1.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SocketIO==4.3.2
flask-talisman==1.0.0
fonttools==4.33.2
ftfy==6.1.1
g2p==0.5.20220318
g2p-en==2.1.0
g2pK==0.9.4
glide-text2im @ git+https://github.com/openai/glide-text2im@69b530740eb6cef69442d6180579ef5ba9ef063e
google-auth==2.6.6
google-auth-oauthlib==0.4.6
googletrans==3.1.0a0
greenlet==1.1.2
grpcio==1.44.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.5.1
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.11.3
inflect==5.5.2
intel-cmplr-lib-rt==2022.1.0
intel-cmplr-lic-rt==2022.1.0
intel-opencl-rt==2022.1.0
intel-openmp==2022.1.0
ipython==7.33.0
itsdangerous==2.1.2
jamo==0.4.1
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
JPype1==1.3.0
kiwisolver==1.4.2
kogpt2-transformers==0.3.0
konlpy==0.6.0
librosa==0.9.1
llvmlite==0.38.0
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mecab-python==1.0.0
mecab-python3==1.0.5
mkl==2022.1.0
mkl-fft==1.3.1
mkl-service==2.4.0
munkres==1.1.4
networkx==2.6.3
nltk==3.7
numba==0.55.1
numpy==1.21.6
oauthlib==3.2.0
openpyxl==3.0.9
packaging==21.3
panphon==0.19.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pooch==1.6.0
prompt-toolkit==3.0.29
protobuf==3.19.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.2
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-engineio==3.14.2
python-mecab-ko==1.0.12
python-socketio==4.6.1
pytz==2022.1
pyworld==0.3.0
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rfc3986==1.5.0
rsa==4.8
sacremoses==0.0.43
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sniffio==1.2.0
SoundFile==0.10.3.post1
SpeechRecognition==3.8.1
stack-data==0.2.0
tbb==2021.6.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
text-unidecode==1.3
tgt==1.4.4
threadpoolctl==3.1.0
tokenizers==0.10.3
torch==1.10.1+cu111
torchaudio==0.10.1+rocm4.1
torchvision==0.11.2+cu111
tqdm==4.64.0
traitlets==5.1.1
transformers==4.11.3
typing_extensions==4.2.0
unicodecsv==0.14.1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.1
wget==3.2
wincertstore==0.2
zipp==3.8.0